requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
starlette==0.37.2
streamlit==1.36.0
python-dotenv==1.0.1
requests==2.32.3
mysql-connector-python==9.0.0
azure-cosmos==4.6.0
jinja2==3.1.4
itsdangerous==2.2.0
openai==1.107.0